gmscompat: Add compatibility layer for unprivileged Google Play services
This is a compatibility layer that allows installing Google Play
services, Google Play Store, and Google Services Framework as regular,
unprivileged apps in the standard app sandbox. Most functionality works
without giving Google Play Services any special privileges, permissions,
or security exceptions, including:

- Account login (including two-factor authentication with NFC security keys)
- Play Store
- Firebase Cloud Messaging notifications (tested with Signal, Discord, Slack, and Gmail)
- Firebase database API (tested with Swift Backup)
- Firebase app indexing
- Google Play Games
- Account settings
- Google My Account
- Autofill
- SMS verification receiver (tested with Signal)
- Play license verification (both in-app purchases and paid apps)
- Play Store app purchases
- Play Store app installation (buggy)
- SafetyNet (basic integrity passes, but not CTS profile checks)
- Dynamite modules (e.g. Maps API v2 and Cronet)
- Play geolocation API (buggy)

The Google Play services family of apps can also be installed in a
secondary user profile for isolation, and all of the above functionality
will still work.

NB: The DropBoxManager and DeviceConfig shims are not functionally
meaningful, but they help reduce log spam that could obscure fatal
errors.

+++ core/java/android/app/NotificationManager.java
@@ -54,6 +54,8 @@ import android.service.notification.ZenPolicy;
 import android.util.Log;
 import android.util.proto.ProtoOutputStream;
 
+import com.android.internal.gmscompat.GmsHooks;
+
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
 import java.util.ArrayList;
@@ -997,6 +999,10 @@ public class NotificationManager {
      * had before it was deleted.
      */
     public void deleteNotificationChannel(String channelId) {
+        if (GmsHooks.skipDeleteNotificationChannel(channelId)) {
+            return;
+        }
+
         INotificationManager service = getService();
         try {
             service.deleteNotificationChannel(mContext.getPackageName(), channelId);
